Extract transformation classes to separate files
