+++ app/src/commonMain/kotlin/diy/lingerie/geometry/transformations/ComplexTransformation.kt
@@ -1,0 +1,11 @@
+package diy.lingerie.geometry.transformations
+
+import diy.lingerie.geometry.Point
+
+sealed class ComplexTransformation : Transformation() {
+    final override fun transform(
+        point: Point,
+    ): Point = simpleTransformations.fold(point) { acc, transformation ->
+        transformation.transform(acc)
+    }
+}
