+++ app/src/commonMain/kotlin/diy/lingerie/geometry/transformations/SimpleTransformation.kt
@@ -1,0 +1,14 @@
+package diy.lingerie.geometry.transformations
+
+sealed class SimpleTransformation : Transformation() {
+    final override val simpleTransformations: List<SimpleTransformation>
+        get() = listOf(this)
+
+    final override fun combineWith(
+        laterTransformations: List<SimpleTransformation>,
+    ): CombinedTransformation = CombinedTransformation(
+        simpleTransformations = listOf(this) + laterTransformations,
+    )
+
+    abstract fun invert(): SimpleTransformation
+}
