+++ app/src/commonMain/kotlin/diy/lingerie/pattern_tool/Outline.kt
@@ -7,11 +7,11 @@ import diy.lingerie.geometry.Line
 import diy.lingerie.geometry.LineSegment
 import diy.lingerie.geometry.Point
 import diy.lingerie.geometry.curves.OpenCurve
-import diy.lingerie.geometry.curves.PrimitiveCurve
 import diy.lingerie.geometry.curves.bezier.BezierCurve
 import diy.lingerie.geometry.splines.ClosedSpline
 import diy.lingerie.geometry.splines.OpenSpline
 import diy.lingerie.geometry.splines.Spline
+import diy.lingerie.geometry.transformations.ReflectionOverLine
 import diy.lingerie.geometry.transformations.Transformation
 import diy.lingerie.utils.iterable.crackAt
 import diy.lingerie.utils.iterable.mapCarrying
@@ -686,7 +686,7 @@ data class Outline(
 
         val mirroredVerges = remainingVerges.map {
             it.reversed().transformBy(
-                transformation = Transformation.ReflectionOverLine(
+                transformation = ReflectionOverLine(
                     line = reflectionLine,
                 ),
             )
