+++ app/src/jvmMain/kotlin/diy/lingerie/simple_dom/svg/SvgShape.kt
